requirements.txt
addict==2.4.0
AHRS==0.3.1
aiofiles==23.2.1
ansi2html==1.8.0
anyio==4.0.0
appnope==0.1.3
argcomplete==3.1.2
asteval==0.9.31
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.3
boto3==1.28.67
botocore==1.31.67
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.36
dacite==1.7.0
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
ffmpeg==1.4
ffmpeg-python==0.2.0
fiftyone==0.23.1
fiftyone-brain==0.14.1
fiftyone-db==0.4.0
fiona==1.9.5
fonttools==4.43.1
ftfy==6.1.1
future==0.18.3
g2o-python==0.0.12
geopandas==0.14.1
glob2==0.7
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hdbscan==0.8.33
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
humanize==4.9.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
ifaddr==0.2.0
imageio==2.31.5
importlib-metadata==6.6.0
inflate64==0.3.1
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kaleido==0.2.1
kiss-icp==0.2.10
kiwisolver==1.4.5
laspy==2.5.1
lazrs==0.5.3
lazy_loader==0.3
lmfit==1.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mongoengine==0.24.2
more-itertools==10.1.0
motor==3.3.1
mrob==0.0.14
multivolumefile==0.2.3
natsort==8.4.0
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==3.2
numpy-quaternion==2022.4.3
open3d==0.17.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
osmnx==1.8.0
ouster-sdk==0.9.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
plyfile==1.0.1
pprintpp==0.4.0
prettytable==3.9.0
priority==2.0.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.6
pybcj==1.0.1
pycpd==2.0.0
pycryptodomex==3.19.0
pydantic==1.10.13
Pygments==2.16.1
pymap3d==3.0.1
pymongo==4.5.0
pyntcloud==0.3.1
pyparsing==3.1.1
pyppmd==1.0.0
pyproj==3.6.1
pyquaternion==0.9.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
pyzstd==0.15.9
rarfile==4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.6
s3transfer==0.7.0
scikit-build==0.17.6
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
shapely==2.0.2
shellingham==1.5.3
simpleicp==2.0.14
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.31.1
strawberry-graphql==0.138.1
tabulate==0.9.0
tenacity==8.2.3
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
trimesh==4.0.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uncertainties==3.1.7
universal-analytics-python3==1.1.1
urllib3==2.0.7
voxel51-eta==0.12.0
wcwidth==0.2.8
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
zeroconf==0.58.2
zipp==3.17.0
